feat(metal): 🎸 Load from file on Apple based environments
